// src/app/models/cliente.ts
export class Cliente {
    _id: number;
    nombre: String;
    nid: String;
    tid: String;
    nombre_contacto: String;
    email_contacto: String;
    direccion: String;
    ciudad: String;
    departamento: String;
    telefonos: any;
    telefono: string;
    logicalerasure: Boolean;
    descripcion: String;
    tipo: number;
    password: String;
    usuarios: any[];
    clientes: any[];
    laboratorios: any[];
    invite: any;
    imagen: String;
    encReporte: String;
    pieReporte: String;
    imagenes: any[];

}